master：管理员菜单功能完善；

+++ ruoyi-common/src/main/java/com/ruoyi/common/constant/Constants.java
@@ -208,7 +208,7 @@ public class Constants
     /**
      * 档口默认权重 0
      */
-    public static final Long STORE_WEIGHT_DEFAULT_ZERO = 0L;
+    public static final Integer STORE_WEIGHT_DEFAULT_ZERO = 0;
 
     public static final String VERSION_LOCK_ERROR_COMMON_MSG = "系统繁忙，请稍后再试";
